integers are now 64 bits

+++ lib.c
@@ -164,7 +164,7 @@ int osc_str_append_bool(struct osc_str *osc_str, int bool)
 	return 0;
 }
 
-int osc_str_append_int(struct osc_str *osc_str, int i)
+int osc_str_append_int(struct osc_str *osc_str, long long int i)
 {
 	int len = osc_str->len;
 	assert(osc_str);
@@ -172,7 +172,7 @@ int osc_str_append_int(struct osc_str *osc_str, int i)
 	osc_str->buf = osc_realloc(osc_str->buf, len + 64);
 	if (!osc_str->buf)
 		return -1;
-	osc_str->len = len + snprintf(osc_str->buf + len, 64, "%d", i);
+	osc_str->len = len + snprintf(osc_str->buf + len, 64, "%lld", i);
 	osc_str->buf[osc_str->len] = 0;
 	return 0;
 }
